% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data_documentation.R
\docType{data}
\name{bathtub}
\alias{bathtub}
\title{Sample of cleaned OSHA accident summaries.}
\format{Corpus object from the \code{tm} package.  Has a meta info of the METHELYNE CHLORIDE labeling called "meth.chl"}
\description{
bathtub consists of several accident reports plus a labeling with a +1 for any report
that had been tagged as related to METHELYNE CHLORIDE.
}
\examples{
library( tm )
data( bathtub )
meta( bathtub, "meth.chl" )
}
\seealso{
Other bathtub: \code{\link{dirtyBathtub}}
}
\concept{bathtub}
\keyword{datasets}
