% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean.text}
\alias{clean.text}
\title{Clean text and get it ready for textreg.}
\usage{
clean.text(bigcorp)
}
\arguments{
\item{bigcorp}{A tm Corpus object.}
}
\description{
Changes multiline documents to single line.  Strips extra whitespace and punctuation.
Changes digits to 'X's.  Non-alpha characters converted to spaces.
}
\examples{
library( tm )
txt = c( "thhis s! and bonkus  4:33pm and Jan 3, 2015. ",
         "   big    space\\n     dawg-ness?")
a <- clean.text( Corpus( VectorSource( txt ) ) )
a[[1]]
}

