% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_phrases.R
\name{cluster.phrases}
\alias{cluster.phrases}
\title{Cluster phrases based on similarity of appearance.}
\usage{
cluster.phrases(result, num.groups = 5, plot = TRUE, yaxt = "n",
  ylab = "", sub = "", main = "Association of Phrases", ...)
}
\arguments{
\item{result}{A similarity matrix from make.similarity.matrix call or an textreg.result object}

\item{num.groups}{Number of groups to box.}

\item{plot}{Actually plot clustering or just calculate it.}

\item{yaxt}{Whether to include a y-axis}

\item{ylab}{Label for y-axis}

\item{sub}{Subtitle for plot}

\item{main}{Title of plot.}

\item{...}{Extra arguments to pass to the plot command.  See par.}
}
\description{
Cluster phrases based on similarity of their appearance in the positive documents.
Can also plot this if so desired.

Uses hclust() with the ``ward.D'' method on 1-S with S from make.similarity.matrix

Warning: for 'negative weight' phrases this method does not do well since it ignores
negative documents.
}
\seealso{
Other Phrase Vizualization: \code{\link{make.appearance.matrix}},
  \code{\link{make.phrase.correlation.chart}},
  \code{\link{make.similarity.matrix}}
}
\concept{Phrase Vizualization}
