% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreg.R
\name{convert.tm.to.character}
\alias{convert.tm.to.character}
\title{Convert tm corpus to vector of strings.}
\usage{
convert.tm.to.character(corpus)
}
\arguments{
\item{corpus}{The tm corpus to convert.}
}
\value{
vector of character.
}
\description{
A utility function useful for testing and some dirty hacks.
This is because the tm package doesn't leave vector corpora of 
strings alone anymore.  

and so sometimes you need to convert your tm object to a string vector
for various reasons, the 
main one being handing it to the C++ method.  It is ugly,
but so it goes.

It is therefore a possibly better decision to pass a filename to a plain-text file
to the textreg call to be loaded by C++ directly.
See \code{\link{textreg}}.
}
