% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreg.R
\name{cpp_build.corpus}
\alias{cpp_build.corpus}
\title{Driver function for the C++ function.}
\usage{
cpp_build.corpus(corpus, labeling, banned = c(), params)
}
\arguments{
\item{corpus}{A list of strings or a corpus from the \code{tm} package.}

\item{labeling}{A vector of +1/-1 or TRUE/FALSE indicating which documents are considered relevant and
which are baseline.  The +1/-1 can contain 0 whcih means drop the document.}

\item{banned}{List of words that should be dropped from consideration.}

\item{params}{List of parameters to pass to the call.}
}
\description{
Given a labeling and a corpus, create a corpus object for use in textreg.
Generally you should use the buildCorpus method, not this method.
}
\details{
Warning: do not call directly.  Use textreg instead
}
\seealso{
textreg, find_C_threshold
}
