% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreg.R
\name{cpp_textreg_driver}
\alias{cpp_textreg_driver}
\title{Driver function for the C++ function.}
\usage{
cpp_textreg_driver(corpus, labeling, banned = c(), params)
}
\arguments{
\item{corpus}{A list of strings or a corpus from the \code{tm} package.}

\item{params}{List of parameters to pass to the call.}
}
\description{
Given a labeling and a corpus, find phrases that predict this labeling.
Generally you should use the textreg method, not this method.
}
\details{
Warning: do not call directly.  Use textreg instead
}
\seealso{
textreg, find_C_threshold
}
