% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data_documentation.R
\docType{data}
\name{dirtyBathtub}
\alias{dirtyBathtub}
\title{Sample of raw-text OSHA accident summaries.}
\format{Dataframe.  Has a meta info of the METHELYNE CHLORIDE labeling, plus 100s of other labels.}
\description{
dirtyBathtub consists of the (more) raw data from which the \code{bathtub} dataset is derived.
}
\examples{
data( dirtyBathtub )
table( dirtyBathtub$fatality )
}
\seealso{
Other bathtub: \code{\link{bathtub}}
}
\concept{bathtub}
\keyword{datasets}
