% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_code.R
\name{find.CV.C}
\alias{find.CV.C}
\title{K-fold cross-validation to determine optimal tuning parameter}
\usage{
find.CV.C(corpus, labeling, banned, K = 5, length.out = 10,
  max_C = NULL, verbose = FALSE, ...)
}
\arguments{
\item{corpus}{The text}

\item{labeling}{The labeling}

\item{banned}{The words to drop.}

\item{K}{Number of folds for K-fold cross-validation}

\item{length.out}{number of values of C to examine from 0 to max_C.}

\item{max_C}{upper bound for tuning parameter; if NULL, sets max_C to threshold C}

\item{verbose}{Print progress}

\item{...}{parameters to be passed to the original textreg() function}
}
\value{
a dataframe containing the mean/standard error of out-of-sample predictions under K-Fold Cross-validation
}
\description{
Given a corpus, divide into K-folds and do test-train spilts averaged over
the folds.
}
\details{
Increments tuning parameter, performs K-fold cross-validation on each C giving a profile
of predictive power for different C.
}
\seealso{
make.CV.chart
}
