% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.fragment.sample}
\alias{fragment.sample}
\alias{is.fragment.sample}
\title{Is object a fragment.sample object?}
\usage{
is.fragment.sample(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a fragment.sample object?
}
\seealso{
Other sample.fragments: \code{\link{print.fragment.sample}};
  \code{\link{sample.fragments}}
}

