% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.textreg.corpus}
\alias{is.textreg.corpus}
\alias{textreg.corpus}
\title{Is object a textreg.corpus object?}
\usage{
is.textreg.corpus(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a textreg.corpus object?
}
\seealso{
Other textreg.corpus: \code{\link{print.textreg.corpus}}
}

