% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.textreg.result}
\alias{is.textreg.result}
\alias{textreg.result}
\title{Is object a textreg.result object?}
\usage{
is.textreg.result(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a textreg.result object?
}
\seealso{
Other textreg.result: \code{\link{print.textreg.result}};
  \code{\link{reformat.textreg.model}}
}

