% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.table.chart}
\alias{list.table.chart}
\title{Graphic showing multiple word lists side-by-side.}
\usage{
list.table.chart(model.list, M = 100, linespace = 4, ytick = NULL,
  dates = NULL, main = paste("Word Appearance for ", attr(model.list,
  "topic"), "\\n(Method: ", attr(model.list, "method"), ")", sep = ""),
  xlab = "Model", mar = c(3, 5, 2.5, 0.1), ...)
}
\arguments{
\item{model.list}{Matrix (or data.frame) from the make.list.table call.}

\item{M}{is the max number of words to show in chart}

\item{linespace}{Where to space}

\item{ytick}{Put y tick marks}

\item{dates}{Dates to put on bottom}

\item{main}{Main title}

\item{xlab}{Label for x-axis}

\item{mar}{Margin of plot (see par)}

\item{...}{Extra arguments for image() call}
}
\description{
This method basically makes a visual plot of a list table (which you call first).
}
\seealso{
make.list.table
}

