% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_phrases.R
\name{make.appearance.matrix}
\alias{make.appearance.matrix}
\title{Make phrase appearance matrix from textreg result.}
\usage{
make.appearance.matrix(result)
}
\arguments{
\item{result}{An textreg.result object.}
}
\value{
A $n X p$ matrix for $n$ documents and $p$ phrases in the result object.  Each entry is a 0/1 value 
   indicating presence of the given phrase in the given document.
}
\description{
Make matrix of which phrases appear in which of the positively
marked documents.
}
\details{
Very similar to phrase.matrix, except this looks only at positively 
marked documents and just returns 1 or 0 on whether any document has
a phrase, rather than giving counts.
This is used by the clustering vizualizations and make.similarity.matrix.
}
\seealso{
make.similarity.matrix

phrase.matrix

Other Phrase Vizualization: \code{\link{cluster.phrases}},
  \code{\link{make.phrase.correlation.chart}},
  \code{\link{make.similarity.matrix}}
}
\concept{Phrase Vizualization}
