% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make.CV.chart}
\alias{make.CV.chart}
\title{Plot K-fold cross validation curves}
\usage{
make.CV.chart(tbl, plot = TRUE, ...)
}
\arguments{
\item{tbl}{Table from find.CV.C}

\item{plot}{TRUE means plot the chart.  False means do not, but return the optimal C}

\item{...}{Parameters to the plot function}
}
\description{
Plot the test error with SE bars for the cross validation.  Also calculate the spot that is 1 SE above the minimum.
Fits the points with loess lines so few points needed in evaluating the function.  All a bit ad hoc and worthy
of improvement.
}
\details{
Not particularly well implemented.
}
\seealso{
find.CV.C
}

