% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceplotter.R
\name{make.path.matrix}
\alias{make.path.matrix}
\title{Generate matrix describing gradient descent path of textreg.}
\usage{
make.path.matrix(res)
}
\arguments{
\item{res}{A textreg.result object.}
}
\description{
Generate a matrix of the sequence of features as they are introduced with the textreg gradient descent 
program along with their coefficients with each step of the descent.
}
\examples{
data( testCorpora )
testI = testCorpora$testI
res = textreg( testI$corpus, testI$labelI, c("frog","goat","bat"), C=2, verbosity=0 )	
make.path.matrix( res )
}
\seealso{
Other plot.path.matrix: \code{\link{path.matrix.chart}},
  \code{\link{plot.textreg.result}}
}
\concept{plot.path.matrix}
