% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_phrases.R
\name{make.phrase.correlation.chart}
\alias{make.phrase.correlation.chart}
\title{Generate visualization of phrase overlap.}
\usage{
make.phrase.correlation.chart(result, count = FALSE, num.groups = 5,
  use.corrplot = FALSE, ...)
}
\arguments{
\item{result}{textreg.result object or a similarity matrix from a make.similarity.matrix call.}

\item{count}{Display counts rather than similarity scores.}

\item{num.groups}{Number of groups to box.}

\item{use.corrplot}{Use the corrplot package of Taiyun Wei (will need to install it).}

\item{...}{Extra arguments to pass to the image() plotting command.  See par.}
}
\description{
Make simple chart showing which phrases have substantial overlap with other phrases.
}
\seealso{
Other Phrase Vizualization: \code{\link{cluster.phrases}},
  \code{\link{make.appearance.matrix}},
  \code{\link{make.similarity.matrix}}
}
\concept{Phrase Vizualization}
