% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_phrases.R
\name{make.similarity.matrix}
\alias{make.similarity.matrix}
\title{Calculate similarity matrix for set of phrases.}
\usage{
make.similarity.matrix(result)
}
\arguments{
\item{result}{An textreg.result object or a matrix from make.appearance.matrix}
}
\description{
First get phrase appearance pattern on positive labeling 
(if not directly passed)
and then calculate similarity matrix of how they are similar
to each other.
}
\details{
Warning: for 'negative weight' phrases this method does not do well since it ignores
negative documents.
}
\seealso{
Other Phrase Vizualization: \code{\link{cluster.phrases}},
  \code{\link{make.appearance.matrix}},
  \code{\link{make.phrase.correlation.chart}}
}
\concept{Phrase Vizualization}
