% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_search_phrases}
\alias{make_search_phrases}
\title{Convert phrases to appropriate search string.}
\usage{
make_search_phrases(phrases)
}
\arguments{
\item{phrases}{List of strings denoting the phrases to be
searched for.}
}
\description{
Will change, e.g., "test * pig+" to appropriate regular
expression to find in the text.
}

