% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceplotter.R
\name{path.matrix.chart}
\alias{path.matrix.chart}
\title{Plot optimization path of textreg.}
\usage{
path.matrix.chart(path.matrix, xlab = "step", ylab = "beta",
  bty = "n", ...)
}
\arguments{
\item{path.matrix}{Either a textreg.result object or a matrix from the make.path.matrix call.}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{bty}{Box for plot}

\item{...}{Arguments to be passed to the matplot() command.}
}
\description{
Plot the sequence of features as they are introduced with the textreg gradient descent 
program.
}
\seealso{
Other plot.path.matrix: \code{\link{make.path.matrix}},
  \code{\link{plot.textreg.result}}
}
\concept{plot.path.matrix}
