% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_searching.R
\name{phrase.count}
\alias{phrase.count}
\title{Count phrase appearance.}
\usage{
phrase.count(phrase, corp)
}
\arguments{
\item{phrase}{A string}

\item{corp}{A corpus object from tm package}
}
\description{
Count number of times a _single_ phrase appears in the corpus
}
\examples{
library( tm )
data( bathtub )
phrase.count( "bathtub", bathtub )
}
\seealso{
Other textregCounting: \code{\link{make.count.table}},
  \code{\link{make.phrase.matrix}}
}
\concept{textregCounting}
