% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_code.R
\name{phrase.matrix}
\alias{phrase.matrix}
\title{Make matrix of where phrases appear in corpus.}
\usage{
phrase.matrix(rules, n)
}
\arguments{
\item{rules}{Either a textreg.result object or the rules list from such an object.}

\item{n}{(Optional) If giving a rules list, the number of documents in corpus.}
}
\description{
Construct a $n X p$ matrix of appearances for selected phrases out of textreg object.
$n$ is the number of documents, $p$ is the number of phrases selected in the result object `rules.'
}
