% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreg.R
\name{phrases}
\alias{phrases}
\title{Get the phrases from the textreg.result object?}
\usage{
phrases(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Get the phrases from the textreg.result object?
}
\seealso{
Other textreg.result: \code{\link{is.textreg.result}},
  \code{\link{print.textreg.result}},
  \code{\link{reformat.textreg.model}}
}
\concept{textreg.result}
