% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceplotter.R
\name{plot.textreg.result}
\alias{plot.textreg.result}
\title{Plot the sequence of features as they are introduced with the textreg gradient descent 
program.}
\usage{
\method{plot}{textreg.result}(x, ...)
}
\arguments{
\item{x}{A textreg.result object.}

\item{...}{Parameters to be passed to path.matrix.chart.}
}
\description{
Simply calls path.matrix.chart.
}
\seealso{
path.matrix.chart

Other plot.path.matrix: \code{\link{make.path.matrix}},
  \code{\link{path.matrix.chart}}
}
\concept{plot.path.matrix}
