% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_searching.R
\name{print.fragment.sample}
\alias{print.fragment.sample}
\title{Pretty print results of phrase sampling object.}
\usage{
\method{print}{fragment.sample}(x, ...)
}
\arguments{
\item{x}{A fragment.sample object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print results of phrase sampling object.
}
\seealso{
Other sample.fragments: \code{\link{is.fragment.sample}},
  \code{\link{sample.fragments}}
}
\concept{sample.fragments}
