% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textreg.R
\name{print.textreg.corpus}
\alias{print.textreg.corpus}
\title{Pretty print textreg corpus object}
\usage{
\method{print}{textreg.corpus}(x, ...)
}
\arguments{
\item{x}{A textreg.corpus object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print textreg corpus object
}
\seealso{
Other textreg.corpus: \code{\link{is.textreg.corpus}},
  \code{\link{textreg.corpus}}
}

