% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreg.R
\name{print.textreg.result}
\alias{print.textreg.result}
\title{Pretty print results of textreg regression.}
\usage{
\method{print}{textreg.result}(x, simple = FALSE, ...)
}
\arguments{
\item{x}{A textreg.result object.}

\item{simple}{TRUE means print out simpler results.  False includes some ugly detail.}

\item{...}{No extra options passed.}
}
\description{
You can also reformat an textreg.result to get simpler diagnostics via \code{\link{reformat.textreg.model}}.
}
\seealso{
reformat.textreg.model

Other textreg.result: \code{\link{is.textreg.result}},
  \code{\link{phrases}},
  \code{\link{reformat.textreg.model}}
}
\concept{textreg.result}
