% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_code.R
\name{refit.model}
\alias{refit.model}
\title{Refit the labels to the selected phrases to get non-shrunk predictions.
(Uses logistic fit rather than linear, as well.)}
\usage{
refit.model(model, method = c("glm", "ols"))
}
\arguments{
\item{model}{The textreg.result object from a text regression}

\item{method}{Either "glm" or "ols".  Type of model to fit.}
}
\value{
A fitted regression model
}
\description{
Refit the labels to the selected phrases to get non-shrunk predictions.
(Uses logistic fit rather than linear, as well.)
}
