% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_code.R
\name{reformat.textreg.model}
\alias{reformat.textreg.model}
\title{Clean up output from textreg.}
\usage{
reformat.textreg.model(model, short = TRUE)
}
\arguments{
\item{model}{The model returned from \code{\link{textreg}}}

\item{short}{True if the output should be abbrviated for easy consumption.}
}
\value{
Dataframe with statistics on the terms in the model
}
\description{
Calculate some useful statistics (percents, etc) and return as dataframe.
}
\seealso{
Other textreg.result: \code{\link{is.textreg.result}},
  \code{\link{phrases}}, \code{\link{print.textreg.result}}
}
\concept{textreg.result}
