% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{save.corpus.to.files}
\alias{save.corpus.to.files}
\title{Save corpus to text (and RData) file.}
\usage{
save.corpus.to.files(bigcorp, filename = "corpus")
}
\arguments{
\item{bigcorp}{A tm Corpus object.}

\item{filename}{The first part of the filename.  A rda and txt extension will be appended to the two generated files.}
}
\description{
Small utility to save a corpus to a text file (and RData file) for ease of use.

It is possibly recommended to pass a filename to the C++ function \code{\link{textreg}}
rather than the entire corpus for
large text since I believe it will otherwise copy over everything due to the coder's (my) poor
understanding of how RCpp converts objects.
}
