% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{testCorpora}
\alias{testCorpora}
\title{Some small, fake test corpora.}
\format{A list of dataframes}
\description{
A list of several fake documents along with some labeling schemes primarily used by the unit testing code.
Also used in some examples.
}
\keyword{datasets}

