% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_searching.R
\name{is.fragment.sample}
\alias{is.fragment.sample}
\alias{fragment.sample}
\title{Is object a fragment.sample object?}
\usage{
is.fragment.sample(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a fragment.sample object?
}
\seealso{
Other sample.fragments: \code{\link{print.fragment.sample}},
  \code{\link{sample.fragments}}
}
\concept{sample.fragments}
