% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_code.R
\name{make.CV.chart}
\alias{make.CV.chart}
\title{Plot K-fold cross validation curves}
\usage{
make.CV.chart(tbl, plot = TRUE, ...)
}
\arguments{
\item{tbl}{Table from find.CV.C}

\item{plot}{TRUE means plot the chart.  False means do not, but return the optimal C}

\item{...}{Parameters to the plot function}
}
\value{
invisible list of the minimum C value and the estimated test error for both the minimum
   and the predicted C corresponding to 1 SE above the minimum estimate.
}
\description{
Make a loess curve with loess() to predict the test error for different values of C by interpolating the passed evaluated
points on the tbl dataframe.
}
\details{
Then plot the test error with SE bars for the cross validation.  Also calculate the spot that is 1 SE above the minimum.
Fits the points with loess lines so, in principle, few actually evaluated points are needed in evaluating the function.  
All a bit ad hoc and worthy of improvement.

Not particularly well implemented.
}
\seealso{
find.CV.C
}
