% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data_documentation.R
\docType{data}
\name{testCorpora}
\alias{testCorpora}
\title{Some small, fake test corpora.}
\format{A list of dataframes}
\description{
A list of several fake documents along with some labeling schemes primarily used by the unit testing code.
Also used in some examples.
}
\keyword{datasets}
